moved some code around
